move constants to dedicated file
